--- src/main/java/com/example/concert_reservation/infra/payment/message/PaymentState.java
@@ -1,5 +1,0 @@
-package com.example.concert_reservation.infra.payment.message;
-
-public enum PaymentState {
-    INIT, POINT, SEAT, FINISHED
-}
