+++ src/main/java/com/example/concert_reservation/presentation/consumer/payment/PaymentMessageConsumer.java
@@ -6,14 +6,12 @@ import com.example.concert_reservation.application.user.UserPointFacade;
 import com.example.concert_reservation.config.exception.CustomException;
 import com.example.concert_reservation.config.exception.CustomExceptionCode;
 import com.example.concert_reservation.domain.common.DataPlatform;
-import com.example.concert_reservation.domain.entity.Concert;
 import com.example.concert_reservation.domain.entity.Reservation;
 import com.example.concert_reservation.domain.entity.User;
 import com.example.concert_reservation.domain.payment.event.PaymentEvent;
 import com.example.concert_reservation.domain.payment.message.PaymentMessage;
 import com.example.concert_reservation.domain.payment.message.PaymentMessageOutboxWriter;
 import com.example.concert_reservation.domain.payment.message.PaymentMessageSender;
-import com.example.concert_reservation.infra.payment.message.PaymentState;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
@@ -43,41 +41,16 @@ public class PaymentMessageConsumer {
     public void update(ConsumerRecord<String, String> record){
         try {
             PaymentMessage paymentMessage = objectMapper.readValue(record.value(), PaymentMessage.class) ;
-            if (paymentMessage.getState() != PaymentState.INIT) {
-                return;
-            }
-            paymentMessage.changeState(PaymentState.POINT);
+//            if (paymentMessage.getState() != PaymentMessage.PaymentState.INIT) {
+//                return;
+//            }
+            paymentMessage.changeState(PaymentMessage.PaymentState.PUBLISHED);
             paymentMessageOutboxWriter.update(paymentMessage);
-            paymentMessageSender.send(paymentMessage);
+//            paymentMessageSender.send(paymentMessage);
         } catch (JsonProcessingException e) {
             throw new CustomException(CustomExceptionCode.INVALID_JSON_TYPE);
         }
 
-    }
-
-    @KafkaListener(topics = "${payment.topic}", groupId = "point-task")
-    @Transactional
-    public void runPointTask(ConsumerRecord<String, String> record) {
-        PaymentMessage paymentMessage = null;
-        PaymentEvent paymentEvent = null;
-        try {
-            paymentMessage = objectMapper.readValue(record.value(), PaymentMessage.class) ;
-            if (paymentMessage.getState() != PaymentState.POINT) {
-                return;
-            }
-            paymentEvent = objectMapper.readValue(paymentMessage.getMessage(), PaymentEvent.class);
-            userPointFacade.usePoint(paymentEvent.getUserId(), paymentEvent.getPrice());
-            paymentMessage.changeState(PaymentState.SEAT);
-            paymentMessageOutboxWriter.update(paymentMessage);
-            paymentMessageSender.send(paymentMessage);
-        } catch (JsonProcessingException e) {
-            log.error("runPointTask json parsing error");
-            throw new CustomException(CustomExceptionCode.INVALID_JSON_TYPE);
-        } catch (Exception e) {
-            e.printStackTrace();
-            log.error("failed point task of payment {}", paymentEvent.getPaymentId());
-            concertFacade.deletePayment(paymentEvent.getPaymentId());
-        }
     }
 
     @KafkaListener(topics = "${payment.topic}", groupId = "reservation-seat-task")
@@ -87,14 +60,11 @@ public class PaymentMessageConsumer {
         PaymentEvent paymentEvent = null;
         try {
             paymentMessage = objectMapper.readValue(record.value(), PaymentMessage.class) ;
-            if (paymentMessage.getState() != PaymentState.SEAT) {
-                return;
-            }
+//            if (paymentMessage.getState() != PaymentMessage.PaymentState.PUBLISHED) {
+//                return;
+//            }
             paymentEvent = objectMapper.readValue(paymentMessage.getMessage(), PaymentEvent.class);
             concertFacade.updateSeatReservationState(paymentEvent.getReservationId());
-            paymentMessage.changeState(PaymentState.FINISHED);
-            paymentMessageOutboxWriter.update(paymentMessage);
-            paymentMessageSender.send(paymentMessage);
         } catch (JsonProcessingException e) {
             log.error("runReservationTask json parsing error");
             throw new CustomException(CustomExceptionCode.INVALID_JSON_TYPE);
@@ -112,9 +82,6 @@ public class PaymentMessageConsumer {
     public void runTokenTask(ConsumerRecord<String, String> record) {
         try {
             PaymentMessage paymentMessage = objectMapper.readValue(record.value(), PaymentMessage.class) ;
-            if (paymentMessage.getState() != PaymentState.FINISHED) {
-                return;
-            }
             PaymentEvent paymentEvent = objectMapper.readValue(paymentMessage.getMessage(), PaymentEvent.class);
             // 실패해도 스케줄러에의해 자동완료
             tokenFacade.expireToken(paymentEvent.getTokenKey());
@@ -128,9 +95,6 @@ public class PaymentMessageConsumer {
     public void runDataPlatFormTask(ConsumerRecord<String, String> record) {
         try {
             PaymentMessage paymentMessage = objectMapper.readValue(record.value(), PaymentMessage.class) ;
-            if (paymentMessage.getState() != PaymentState.FINISHED) {
-                return;
-            }
             PaymentEvent paymentEvent = objectMapper.readValue(paymentMessage.getMessage(), PaymentEvent.class);
             String text = "success";
             Reservation reservation = concertFacade.getReservation(paymentEvent.getReservationId());
